# Try replacing pow with exp log