I may try to add Float32 perf tests, just to see / verify from here that this is problematic.