@charleskawczynski 2022-08-17
Clear improvement here.

```
# Main:
[ Info: Microbenchmarks for  Float64                                              Float32
┌───────────────────────────┬────────────┬────────────┬────────────┬────────────┬┬────────────┬────────────┬────────────┬────────────┐
│ Constructor               │       Time │       Time │       Time │       Time ││       Time │       Time │       Time │       Time │
│ (+conditions)             │        min │        max │       mean │     median ││        min │        max │       mean │     median │
├───────────────────────────┼────────────┼────────────┼────────────┼────────────┼┼────────────┼────────────┼────────────┼────────────┤
│ PhaseEquil_ρeq_dry        │ 289.175 ns │ 433.540 ns │ 291.341 ns │ 290.175 ns ││ 267.240 ns │ 956.055 ns │ 269.206 ns │ 267.905 ns │
│ PhaseEquil_ρeq_sat_adjust │   1.524 μs │   2.284 μs │   1.536 μs │   1.530 μs ││   1.389 μs │   2.999 μs │   1.403 μs │   1.398 μs │
│ PhaseEquil_ρTq_dry        │ 157.477 ns │ 216.533 ns │ 159.794 ns │ 159.246 ns ││ 144.394 ns │ 207.228 ns │ 151.155 ns │ 149.983 ns │
│ PhaseEquil_ρTq_moist      │ 157.887 ns │   1.725 μs │ 159.969 ns │ 158.906 ns ││ 139.740 ns │ 216.925 ns │ 153.155 ns │ 149.488 ns │
│ PhaseEquil_pTq_dry        │ 164.817 ns │ 227.485 ns │ 165.749 ns │ 165.111 ns ││ 140.627 ns │ 215.226 ns │ 153.082 ns │ 150.441 ns │
│ PhaseEquil_pTq_moist      │ 164.817 ns │ 241.832 ns │ 165.742 ns │ 165.091 ns ││ 140.517 ns │ 417.333 ns │ 153.213 ns │ 150.377 ns │
│ PhaseEquil_peq_dry        │ 296.778 ns │   6.089 μs │ 299.130 ns │ 297.314 ns ││ 252.400 ns │ 401.085 ns │ 275.309 ns │ 270.397 ns │
│ PhaseEquil_peq_sat_adjust │   5.977 μs │  13.505 μs │   6.012 μs │   5.988 μs ││   5.182 μs │  12.304 μs │   5.647 μs │   5.376 μs │
│ PhaseEquil_phq_dry        │ 303.268 ns │   1.153 μs │ 305.047 ns │ 303.787 ns ││ 257.304 ns │ 406.046 ns │ 280.188 ns │ 267.994 ns │
│ PhaseEquil_phq_sat_adjust │   5.013 μs │   9.031 μs │   5.042 μs │   5.021 μs ││ 981.700 ns │   4.004 μs │   1.080 μs │   1.018 μs │
│ PhaseEquil_ρθq_dry        │ 414.160 ns │ 560.565 ns │ 416.855 ns │ 415.415 ns ││ 282.561 ns │ 500.521 ns │ 302.190 ns │ 292.557 ns │
│ PhaseEquil_ρθq_sat_adjust │   8.962 μs │  17.247 μs │   9.021 μs │   8.974 μs ││   7.241 μs │  25.698 μs │   8.393 μs │   7.978 μs │
│ PhaseEquil_pθq_dry        │ 394.350 ns │ 560.285 ns │ 410.719 ns │ 408.365 ns ││ 273.534 ns │ 394.033 ns │ 297.970 ns │ 292.652 ns │
│ PhaseEquil_pθq_sat_adjust │   2.173 μs │  24.511 μs │   2.268 μs │   2.252 μs ││   1.924 μs │   6.075 μs │   2.107 μs │   1.992 μs │
│ PhaseEquil_ρpq_dry        │  16.749 ns │  48.090 ns │  17.676 ns │  16.791 ns ││  15.698 ns │  76.518 ns │  16.674 ns │  15.729 ns │
│ PhaseEquil_ρpq_sat_adjust │  16.212 ns │  42.163 ns │  17.484 ns │  17.354 ns ││  15.192 ns │  46.164 ns │  16.511 ns │  15.730 ns │
└───────────────────────────┴────────────┴────────────┴────────────┴────────────┴┴────────────┴────────────┴────────────┴────────────┘
# This PR:
[ Info: Microbenchmarks for  Float64                                              Float32
┌───────────────────────────┬────────────┬────────────┬────────────┬────────────┬┬────────────┬────────────┬────────────┬────────────┐
│ Constructor               │       Time │       Time │       Time │       Time ││       Time │       Time │       Time │       Time │
│ (+conditions)             │        min │        max │       mean │     median ││        min │        max │       mean │     median │
├───────────────────────────┼────────────┼────────────┼────────────┼────────────┼┼────────────┼────────────┼────────────┼────────────┤
│ PhaseEquil_ρeq_dry        │ 124.070 ns │ 167.144 ns │ 124.947 ns │ 124.435 ns ││ 124.235 ns │ 167.560 ns │ 126.254 ns │ 125.834 ns │
│ PhaseEquil_ρeq_sat_adjust │ 718.803 ns │  11.893 μs │ 725.676 ns │ 721.825 ns ││ 690.167 ns │   2.070 μs │ 694.573 ns │ 691.333 ns │
│ PhaseEquil_ρTq_dry        │  75.210 ns │ 109.864 ns │  76.247 ns │  75.930 ns ││  74.943 ns │ 330.365 ns │  76.306 ns │  75.288 ns │
│ PhaseEquil_ρTq_moist      │  71.770 ns │   1.353 μs │  73.460 ns │  72.813 ns ││  73.841 ns │ 280.077 ns │  77.781 ns │  77.588 ns │
│ PhaseEquil_pTq_dry        │  73.717 ns │ 112.684 ns │  74.238 ns │  73.954 ns ││  68.956 ns │ 110.840 ns │  71.787 ns │  71.507 ns │
│ PhaseEquil_pTq_moist      │  73.057 ns │ 112.031 ns │  73.698 ns │  73.405 ns ││  68.952 ns │ 105.345 ns │  71.397 ns │  71.438 ns │
│ PhaseEquil_peq_dry        │ 129.374 ns │ 171.570 ns │ 130.291 ns │ 129.804 ns ││ 110.272 ns │ 183.160 ns │ 123.942 ns │ 121.780 ns │
│ PhaseEquil_peq_sat_adjust │   3.078 μs │   5.538 μs │   3.100 μs │   3.086 μs ││   2.846 μs │   6.219 μs │   3.007 μs │   2.971 μs │
│ PhaseEquil_phq_dry        │ 141.701 ns │ 205.215 ns │ 142.568 ns │ 142.034 ns ││ 134.458 ns │   1.958 μs │ 143.609 ns │ 141.044 ns │
│ PhaseEquil_phq_sat_adjust │   2.596 μs │  25.453 μs │   2.617 μs │   2.604 μs ││ 501.117 ns │ 803.638 ns │ 565.538 ns │ 552.904 ns │
│ PhaseEquil_ρθq_dry        │ 251.892 ns │ 323.304 ns │ 253.719 ns │ 252.707 ns ││ 144.463 ns │ 208.552 ns │ 153.859 ns │ 154.564 ns │
│ PhaseEquil_ρθq_sat_adjust │   6.011 μs │  14.942 μs │   6.421 μs │   6.433 μs ││   5.636 μs │   9.941 μs │   5.907 μs │   5.835 μs │
│ PhaseEquil_pθq_dry        │ 236.379 ns │ 376.877 ns │ 246.372 ns │ 245.578 ns ││ 140.181 ns │ 204.420 ns │ 147.256 ns │ 145.324 ns │
│ PhaseEquil_pθq_sat_adjust │   1.662 μs │   4.488 μs │   1.675 μs │   1.666 μs ││   1.512 μs │   5.200 μs │   1.570 μs │   1.568 μs │
│ PhaseEquil_ρpq_dry        │  19.838 ns │  37.349 ns │  20.025 ns │  19.959 ns ││  15.697 ns │  36.021 ns │  16.011 ns │  15.721 ns │
│ PhaseEquil_ρpq_sat_adjust │  16.751 ns │  41.749 ns │  17.374 ns │  17.339 ns ││  15.192 ns │  40.209 ns │  15.975 ns │  15.728 ns │
└───────────────────────────┴────────────┴────────────┴────────────┴────────────┴┴────────────┴────────────┴────────────┴────────────┘
```